Review gate: compensate on confirm failure, fail-fast timeouts, tighten tests (#52)
Findings from the SDD review gate (riviera-review-overlay + code-review):
- Confirm-after-commit asymmetry: the Succeeded/stub path confirmed after
  the reserve commit with no compensation, so a confirm failure would
  strand the set AWAITING_PAYMENT (default profile has no TTL sweep). Now
  compensates symmetrically with the Failed branch (release + rethrow);
  pinned by CreateBookingServiceTest.compensatesWhenConfirmFailsAfterCommit.
- Timeout overflow: (int) Duration.toMillis() -> Math.toIntExact so a
  misconfigured > ~24.8-day timeout fails fast instead of wrapping negative.
- Tests: corrected persistsBeforePayingThenAwaits to claim only the
  insert->pay ordering it proves; CreateBookingPaymentFailureIT now asserts
  exactly one booking row for the (set, date).

+++ platform/src/main/java/ai/riviera/platform/payment/infrastructure/StripeConfig.java
@@ -38,7 +38,9 @@ class StripeConfig {
 	static StripeClient.StripeClientBuilder clientBuilder(StripeProperties properties) {
 		return StripeClient.builder()
 				.setApiKey(properties.apiKey())
-				.setConnectTimeout((int) properties.connectTimeout().toMillis())
-				.setReadTimeout((int) properties.readTimeout().toMillis());
+				// toIntExact fails fast on a misconfigured > ~24.8-day timeout rather than silently
+				// wrapping to a negative millisecond value.
+				.setConnectTimeout(Math.toIntExact(properties.connectTimeout().toMillis()))
+				.setReadTimeout(Math.toIntExact(properties.readTimeout().toMillis()));
 	}
 }
